+++ src/pages/Main/Card.jsx
@@ -5,13 +5,6 @@ import { useTheme } from "@mui/material";
 import PeopleOutlinedIcon from "@mui/icons-material/PeopleOutlined";
 import StatBox from './StatBox';
 import ProgressCircle from "./ProgressCircle";
-import { toast, ToastContainer } from "react-toastify";
-import "react-toastify/dist/ReactToastify.css";
-
-
-
-
-
 
 const List = () => {
 
@@ -21,47 +14,9 @@ const List = () => {
     const colors = tokens(theme.palette.mode);
 
 
-    const notify = () => {
-        if (1 > 0) {
-            toast(`스크리닝 환자가 ${todayScreening}명 추가 되었습니다`,
-                { autoClose: 3000 } //3초
-            );
-        }
-    }
-
-    // toast 컴포넌트
-    useEffect(() => {
-        const interval = setInterval(() => {
-            notify();
-
-        }, 
-        10000
-        ); // 한 시간(밀리초 단위)
-        // 컴포넌트가 언마운트 될 때 타이머를 정리합니다.
-        return () => clearInterval(interval);
-    }, []); // []를 전달하여 한 번만 실행되도록 설정합니다.
-
     return (
         <>
             <div>
-            <ToastContainer
-                    position="top-right"
-                    autoClose={30000}
-                    hideProgressBar={false}
-                    newestOnTop={true}
-                    closeOnClick={true}
-                    pauseOnHover={true}
-                    draggable={true}
-                    progressStyle={{ background: "#3e4396" }}
-                    bodyStyle={{
-                        fontFamily: "Arial, sans-serif",
-                        fontSize: "14px",
-                        fontWeight: "bold",
-                        color: "#000000",
-                        backgroundColor: "#ffffff",
-                    }}
-                />
-
                 <Box
                     display="grid"
                     gridTemplateColumns="repeat(12, 1fr)"
